refactor: Split restore into per-generation strategies
RestoreDeviceConfig keeps orchestration only: backup lookup, identity and generation checks, selection and ordering, unknown-key surfacing, aggregation, and the reboot-if-applied rule. The Gen2 path (SetConfig/PutCode/Schedule.Replace, read-only strip) and the Gen1 path (raw-settings replay, mode pre-phase, WiFi fan-out, legacy reboot) move unchanged into ComponentRestoreStrategy implementations under use_cases/restore_strategies.

Strategies are constructed per restore run from the injected gateway: the Gen1 strategy holds the loaded settings between prepare and the component loop, and the use case is a container singleton, so a shared strategy instance would leak state across concurrent restores. The public constructor keeps its signature; the mode timing kwargs are forwarded to the Gen1 strategy.
